backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via
    #   httpx
    #   openai
blinker==1.9.0
    # via flask
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via flask
distro==1.9.0
    # via openai
flask==3.1.0
    # via allergy-checker (pyproject.toml)
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.0
    # via openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
jiter==0.8.0
    # via openai
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
openai==1.56.1
    # via allergy-checker (pyproject.toml)
pydantic==2.10.3
    # via openai
pydantic-core==2.27.1
    # via pydantic
python-dotenv==1.0.1
    # via allergy-checker (pyproject.toml)
requests==2.32.3
    # via allergy-checker (pyproject.toml)
sniffio==1.3.1
    # via
    #   anyio
    #   openai
tqdm==4.67.1
    # via openai
typing-extensions==4.12.2
    # via
    #   openai
    #   pydantic
    #   pydantic-core
urllib3==2.2.3
    # via requests
werkzeug==3.1.3
    # via flask
